requirements.txt
Django==2.1.5
fastai==1.0.42
Pillow==5.4.1
psycopg2==2.7.7
gunicorn
